requirements.txt
python-vlc
mysql-connector